More secure handling of filenames.
Try to avoid being tricked into sending files we shouldn't, or asking
nytprofhtml to write to files it shouldn't be.

For instance, with a cleverly-crafted call to /nytprof/html/... using
combinations of ".." and %2F, it would be possible to access other files.

(This is why the documentation contains a big disclaimer :) )