Give Devel::NYTProf a temp filename to write to.
Devel::NYTProf seems to insist on opening a profile file to write, even if
you've told it not to start yet with start=no.

Get a temporary filename, and let it open that.  It's a throwaway file that
shouldn't actually be written to; when we actually tell Devel::NYTProf to start
profiling, we'll give it a new file name to write to in the desired target
directory.

This is nasty and hackish, but should make it work for situations where the
appdir isn't writable.

I'll raise a bug against Devel::NYTProf, as this behaviour doesn't seem to be
desirable or expected.