Merging in CPAN version 0.31, unfucking history.
Somehow, I really fucked this up; the current version on CPAN contains code that
isn't in the history here in this repo.  I can only guess that I worked on it
from two different checkouts (probably my home server and a dev VM at work, or
locally on my laptop) and, for some stupid reason, ended up force-pushing from a
checkout that was behind, trashing history.

However I managed to do it, this merges in the changes made in master since CPAN
release 0.31 with the stuff in that release that's not in master.

Note: this will introduce the problem in Issue #14 which will need fixing before
this hits CPAN.