Add profiling_enabled for more fine-grained control.
Allows you to disable profiling whilst still serving up the previous profiled
runs for your perusal.  It should be possible to enable/disable profiling on the
fly by altering the config at runtime.

The existing C<enabled> setting, added in the last release, acts as an overall
enabled or not switch - if it's disabled in the config, then the routes and
hooks don't get installed at all.  If it was enabled in the config but later
disabled at runtime, then the routes will do what you'd probably expect by
refusing to run - but they'll still get called.  However, if it was not enabled
in the config, you can't then enable it later at runtime, as the hooks & routes
didn't get installed.